// src/alram-push/dto/berth.dto.ts
export class BerthPyDto {
  oid: string;
  trminlCode?: string;
  berthCode?: string;
  trminlVoyg?: string;
  wtorcmpCode?: string;
  trminlShipnm?: string;
  shipRute?: string;
  csdhpPrarnde?: string;
  tkoffPrarnde?: string;
  csdhpDrc?: string;
  workStarDay?: string;
  workFiniDay?: string;
  carryFiniDay?: string;
  landngQy?: string;
  shipngQy?: string;
  reshmtQy?: string;
  predBerth?: string;
  shipment?: string;
  shifting?: string;
  createDate?: Date;
  updateDate?: Date;
}